Remove password fields for now
The password is really used as a key for client side encryption and
shouldn't appear to be optional. We don't want the user to pick a weak
password either.
So it's better to implement a separate *download* password later.

We now generate a password with more bits and pass it as b64 encoded
string in the fragment.

This invalidates all existing download links.

On the bright side this removes the crypto pbkdf2 dependency.